+++ mitdevices/l6810a.c
@@ -3,6 +3,8 @@
 #include <mitdevices_msg.h>
 #include <mds_stdarg.h>
 #include "l6810a_gen.h"
+#include "devroutines.h"
+
 extern int CamXandQ();
 
 #define min(a,b) ((a) < (b)) ? (a) : (b)
@@ -147,7 +149,7 @@ int l6810a___store(struct descriptor *niddsc_ptr, InStoreStruct *in_struct)
 #define pio(f,a,d,q)  return_on_error(DevCamChk(CamPiow(in_struct->name, a, f, d, 16, 0), &one, &q))
 #define CHAN_NID(chan, field) c_nids[L6810A_N_CHANNELS+chan*L6810A_K_NODES_PER_CHANNEL+field]
 
-  static unsigned long one = 1;
+  static int one = 1;
   static int c_nids[L6810A_K_CONG_NODES];
 
   static DESCRIPTOR_A_BOUNDS(raw,sizeof(short),DTYPE_W,0,1,0);
